requirements.txt
git+https://github.com/Rapptz/discord.py
PyNaCl==1.4.0
pandas
dnspython==2.0.0
async-timeout==3.0.1
praw
wikipedia
lavalink
wavelink
discord.py[voice]==1.5.1
websockets
youtube_dl
colorlog
aiohttp
lyricsgenius
azapi
requests
html